# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Secure MOM",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/meetings/{ident}/snapshots": {
      "post": {
        "summary": "Snapshot",
        "operationId": "snapshot_api_v1_meetings__ident__snapshots_post",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Revision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Snapshots",
        "operationId": "list_snapshots_api_v1_meetings__ident__snapshots_get",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/snapshots/{ident}/approve": {
      "post": {
        "summary": "Approve",
        "operationId": "approve_api_v1_snapshots__ident__approve_post",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Revision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recipient-groups": {
      "get": {
        "summary": "Groups",
        "operationId": "groups_api_v1_recipient_groups_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Group",
        "operationId": "group_api_v1_recipient_groups_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Group"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/snapshots/{ident}/exports/{format}": {
      "get": {
        "summary": "Export",
        "operationId": "export_api_v1_snapshots__ident__exports__format__get",
        "parameters": [
          {
            "name": "ident",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ident"
            }
          },
          {
            "name": "format",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Format"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/glossary": {
      "get": {
        "summary": "Get Glossary",
        "operationId": "get_glossary_api_v1_settings_glossary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "put": {
        "summary": "Glossary",
        "operationId": "glossary_api_v1_settings_glossary_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Glossary"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Glossary": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "terms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Terms"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "terms"
        ],
        "title": "Glossary"
      },
...
      "Group": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 0
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "addresses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Addresses"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "addresses"
        ],
        "title": "Group"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Revision": {
        "properties": {
          "revision": {
            "type": "integer",
            "title": "Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision"
        ],
        "title": "Revision"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}